Move fast merging to atlannot.merge
